파일 수정 db서버 api서버 병합
